restructure gui js directory
